+++ src/models/Data.java
@@ -2,9 +2,9 @@ package models;
 
 public class Data {
     public double[][] X; // 2D points
-    public double[] y; // class labels
+    public int[] y; // class labels
 
-    public Data(double[][] X, double[] y) {
+    public Data(double[][] X, int[] y) {
         this.X = X;
         this.y = y;
     }
